wo #5532 . refactor lp47 for use in lp57
